pocetak

+++ demo/src/main/java/com/example/demo/repositories/IProductRepository.java
@@ -4,6 +4,7 @@ import com.example.demo.entities.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
 @Repository
-public interface IProductRepository extends JpaRepository<Product, Long> {
+public interface IProductRepository extends JpaRepository<Product, Integer> {
 }
